allow partial hours

+++ TimeCardr/Entry.cs
@@ -8,9 +8,9 @@ namespace TimeCardr
 		public string ResourceName { get; private set; }
 		public string Project { get; private set; }
 		public string Task { get; private set; }
-		public int Hours { get; private set; }
+		public double Hours { get; private set; }
 
-		public Entry(DateTime date, string resourceName, string project, string task, int hours)
+		public Entry(DateTime date, string resourceName, string project, string task, double hours)
 		{
 			Date = date;
 			ResourceName = resourceName;
